backend/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
cryptography==42.0.2

# Configuration
pyyaml==6.0.1

# HTTP client (for Discord webhooks, SMTP)
httpx==0.26.0
aiosmtplib==3.0.1

# Email templates
jinja2==3.1.3

# HTTPS / Let's Encrypt
certbot==2.8.0

# Utilities
python-dateutil==2.8.2
